Refactor LastFuncNum related properties
